feat: 1st Draft of `EmailPassword` login with Next.js App Dir
There are still some things that we don't like... such as:
- The need to import the SuperTokens `initialize.ts` file
  anywhere that its utilities are needed. (Not sure why
  this happens).
- Some weird workarounds that are needed for using Next.js
  properly (like relying on headers to access global server
  data instead of relying on a feature native to the framework
  since Next doesn't have a solution to this problem).

Nonetheless, our Playwright tests are passing for our Next.js
app. So... things seem to be in good shape. And the amount
of brute forcing necessary to get everything going is surprisingly
lower than one would expect with a typical Next app.

Note: `instrumentation.ts` doesn't yet do anything meaningful
because of the `import` issue described earlier. But we're keeping
the file there as a placeholder to remind us to address the issue
if possible sometime in the future.