refactor: Make App Router and Pages Router independent, templated apps
The reason for this change was simplicity. By forcing the App Router
and the Pages Router to share code, we ended up with odd conventions
that would never appear in a realworld application. For instance,
in the real world, most `Header` components would not need to know
whether they were being used in an App Router vs. a Pages Router.
Similarly, as far as the Pages Router is concerned, a developer
would define the `/api/email-exists` route in a `/pages/api/email-exists.ts`
file. And in that case, they could use Next's native handlers
instead of relying on our clever technique. Things get even more
compliated when cookie management gets involves.

We want developers to be able to see what a REALISTIC App Router
would look like and what a REALISTIC Pages Router would look like.
And the more that we work on this project, the more it becomes
apparent that the only way to accomplish this is to create 2 separate
apps that don't need to share code with each other in an awkward
fashion. Thus, we've split the routers into separate templates.

Since we aren't interested in maintaining duplicates of files, we're
also using symbolic links. Whenever the Pages Router uses a file
that has the exact same content as something in the App Router, we
just create a symbolic link from the Pages Router to the App Router.
There's a little bit of boilerplate necessary to make this work
(see the `/pages-router/tsconfig.json`), but ultimately it's simpler
than trying to keep duplicates of files up to date with each other.

Admittedly, this approach adds its own amount of complexity to our
repo -- far more than we would have liked. But I think it's just
what has to be done to support the two very different kinds of routers
that Next.js has. Although Remix also has major version upgrades,
it doesn't require anything like this.

An NPM script was needed to automatically pull from the templates
to create a `src` directory that can run a Next.js application.
This auto-genrated `src` directory is `gitignore`d for obvious reasons.