chore: Move comment about Server Action Form Resets to Private Page
We now force Server Actions to be called strictly from the client's side
on our pages that have client-side validation. Consequently, the issue
of automatic form resets is no longer a concern on those pages. (Manual
invocations of Server Actions do not seem to produce automatic form
resets.)

We don't use this special approach on the Private Page, meaning that we
meaning that we still have to deal with automatic form resets on that
page. So we need to address this problem in the future. Now that this
problem is just restricted to the Private Page, it's much less of a
priority. However, this behavior should probably still be documented
in the README.