feat: Get Playwright tests working with AND without JavaScript enabled
We documented our findings in `NOTES.md` at:
ITenthusiasm/remix-supertokens@2726466dd8195f16b068d561f05dfa40f7afec5a

NOTE: Due to limitations in Next's App Router, there are some tests that
won't work. (The Pages Router does not have any problems.) Specifically,
test will break whenever a user receives a 303 response as a result of BOTH
1) Having JavaScript enabled AND 2) Submitting a form using Server Actions.
In these scenarios, Next.js fails to navigate to the correct page;
it simply does nothing. (This can happen, for example, if a user tries
to submit a Server Action form on a secured page when their access
token is expired or invalid.)

Other frameworks like Remix do not have this problem. The Next.js
team might not be aware of this issue. So it could be good to raise
it on GitHub at some point.